Initial javaviz layout based on gui layout.
This is true until the user saves a layout in javaviz. Then that layout
is always used as the initial layout (until they save again, of course)
The one exception is when the top level network doesn't have a label, in
that case it is currently impossible to save a javaviz layout because
the name of the network is different everytime javaviz runs.

Fixes #57.