Implemented setup.py
Going all out with setuptools. Now, you can do the usual
python setup.py install or develop, and it will create a
nengo_gui script that you can run from anywhere on the command
line. Not sure how this works with getting an icon for Windows,
if that's critical I can play around with it.

Other changes
- Added version.py to track version number; v0.1.0-dev right now
- Moved main to nengo_gui, put in function so it works with setuptools
- Make README an RST doc as that's what PyPI can parse
- Added myself to CONTRIBUTORS.md